# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NH Api",
        "version": "1"
    },
    "paths": {
...
        "/api/orders-list/{outlet_id}": {
            "get": {
                "tags": [
                    "Outlet Orders"
                ],
                "summary": "Get outlet orders list",
                "operationId": "672b144754dbd5c7addc81e9e04d9d93",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorize Access, Invalid Token or Token has expired",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/orders-list/{outlet_id}/{status}": {
            "get": {
                "tags": [
                    "Outlet Orders"
                ],
                "summary": "Get outlet orders list by status",
                "operationId": "8fa117a852b06a6ef520a62eee003454",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorize Access, Invalid Token or Token has expired",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/orders-details/{outlet_id}/{order_id}": {
            "get": {
                "tags": [
                    "Outlet Orders"
                ],
                "summary": "Get order details by outlet id and order id",
                "operationId": "2ef784623c431f04e2a5d89d20941234",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorize Access, Invalid Token or Token has expired",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/order-process": {
            "post": {
                "tags": [
                    "Outlet Orders"
                ],
                "summary": "Update order",
                "description": "Returns project data",
                "operationId": "update",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "outlet_id",
                                    "status"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": "39753"
                                    },
                                    "outlet_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "status": {
                                        "type": "integer",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Order updated successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/payment": {
            "post": {
                "tags": [
                    "Outlet Payment"
                ],
                "summary": "Receive Online payment",
                "description": "Receive online payment when customer comes to the outlet for order pickup",
                "operationId": "payment",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "order_id",
                                    "outlet_id",
                                    "transaction_amount",
                                    "payer_name",
                                    "payer_phone"
                                ],
                                "properties": {
                                    "order_id": {
                                        "type": "integer",
                                        "example": "39753"
                                    },
                                    "outlet_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "transaction_amount": {
                                        "type": "number",
                                        "format": "double",
                                        "example": "5156.25"
                                    },
                                    "payer_name": {
                                        "type": "string",
                                        "example": "Md Frank"
                                    },
                                    "payer_phone": {
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "note": {
                                        "type": "string",
                                        "example": "something"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "http",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}